Not all changes were staged in the last commit

+++ Vitro/src/Vitro/API/DirectX/IndexBuffer.cpp
@@ -6,15 +6,15 @@
 namespace Vitro::DirectX
 {
 	IndexBuffer::IndexBuffer(const Array<uint32_t>& indices) :
-		IndexBuffer(indices.Raw(), indices.Count())
+		IndexBuffer(indices.Raw(), static_cast<uint32_t>(indices.Count()))
 	{}
 
-	IndexBuffer::IndexBuffer(const uint32_t indices[], size_t count)
+	IndexBuffer::IndexBuffer(const uint32_t indices[], uint32_t count)
 	{
 		D3D11_BUFFER_DESC bd{0};
-		bd.ByteWidth = static_cast<UINT>(count * sizeof(uint32_t));
-		bd.Usage = D3D11_USAGE_DEFAULT;
-		bd.BindFlags = D3D11_BIND_INDEX_BUFFER;
+		bd.ByteWidth	= count * sizeof(uint32_t);
+		bd.Usage		= D3D11_USAGE_DEFAULT;
+		bd.BindFlags	= D3D11_BIND_INDEX_BUFFER;
 
 		D3D11_SUBRESOURCE_DATA srd{0};
 		srd.pSysMem = indices;
